Update index.md
Too many commas. Perhaps the intent is to make it read like live speech, but it isn't. It is static words on a web page.

Is there a link to a blog posting talking about 'parasitic intermediaries'? This phrase really jumps out for its intensity, but if it was explained it would help a potential user make the decision to use Status.

At one point "The Ethereum blockchain" is mentioned, and then later under **Security** "Blockchain" is again mentioned. Are these the same blockchain? If so, why is one capitalized and the other not? Could the same terms be used in both cases for uniformity? Can we provide a link for both to whatever the "The Ethereum blockchain" actually is?

Also, "...the Blockchain, it handles crypto-tokens that have real value and allows 3rd party code execution." In this phrase what handles crypto-tokens and allows 3rd party code execution? Blockchain, or Status? My change presumes that Status is what is referred to.

I also suggest fewer contractions. The site overall aims for serious goals, but all the 'll and 'nt 're detract from the seriousness. The two exceptions would be "It's bug hunting season!" and "It's all about sex appeal!" since these phrases are clearly intended to be light hearted. To be honest I almost decided to not get involved because something like 'sex appeal' seems out of place in this day and age, which is why I previously suggested "visual appeal".